add different rendering option

+++ src/triangle.c
@@ -34,7 +34,7 @@ void draw_filled_triangle(int x0, int y0, int x1, int y1, int x2, int y2,
     fill_flat_top_triangle(x0, y0, x1, y1, x2, y2, color);
   } else {
     int My = y1;
-    int Mx = ((float)((x2 - x0) * (y1 - y0)) / (float)(y2 - y0)) + x0;
+    int Mx = (((x2 - x0) * (y1 - y0)) / (y2 - y0)) + x0;
 
     // Draw flat-bottom triangle
     fill_flat_bottom_triangle(x0, y0, x1, y1, Mx, My, color);
@@ -46,12 +46,15 @@ void draw_filled_triangle(int x0, int y0, int x1, int y1, int x2, int y2,
 
 void fill_flat_bottom_triangle(int x0, int y0, int x1, int y1, int x2, int y2,
                                uint32_t color) {
+  // Find the two slopes (two triangle legs)
   float inv_slope_1 = (float)(x1 - x0) / (y1 - y0);
   float inv_slope_2 = (float)(x2 - x0) / (y2 - y0);
 
+  // Start x_start and x_end from the top vertex (x0,y0)
   float x_start = x0;
   float x_end = x0;
 
+  // Loop all the scanlines from top to bottom
   for (int y = y0; y <= y2; y++) {
     draw_line(x_start, y, x_end, y, color);
     x_start += inv_slope_1;
@@ -61,12 +64,15 @@ void fill_flat_bottom_triangle(int x0, int y0, int x1, int y1, int x2, int y2,
 
 void fill_flat_top_triangle(int x0, int y0, int x1, int y1, int x2, int y2,
                             uint32_t color) {
+  // Find the two slopes (two triangle legs)
   float inv_slope_1 = (float)(x2 - x0) / (y2 - y0);
   float inv_slope_2 = (float)(x2 - x1) / (y2 - y1);
 
+  // Start x_start and x_end from the bottom vertex (x2,y2)
   float x_start = x2;
   float x_end = x2;
 
+  // Loop all the scanlines from bottom to top
   for (int y = y2; y >= y0; y--) {
     draw_line(x_start, y, x_end, y, color);
     x_start -= inv_slope_1;
